+++ ligmos/workers/__init__.py
@@ -1,3 +1,4 @@
 from . import connSetup
+from . import confUtils
 from . import workerSetup
 from . import defaultParser
